pyproject.toml
[project]
name = "mongodb-tokenizer"
version = "0.1.0"
description = ""
authors = [
    {name = "Evgeniy Trubnikov",email = "kiwibirdparmesan@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "pymongo (>=4.15.1,<5.0.0)",
    "beanie (>=2.0.0,<3.0.0)",
    "spacy (>=3.8.7,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [
    { include = "src" }
]

[tool.poetry.scripts]
app = "src.main:run"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
